Merged develop to test (part of test deployment).

+++ src/backend/DocDownloaderAPI/src/main/java/ca/bc/gov/ag/courts/model/Job.java
@@ -55,7 +55,7 @@ public class Job implements Serializable {
 	@JsonProperty("endDeliveryDtm")
 	private String endDeliveryDtm;
 	@JsonProperty("percentageComplete")
-	private Integer percentageComplete;
+	private float percentageComplete;
 	@JsonProperty("ordsFileName")
 	private String ordsFileName;
 	@JsonProperty("fileName")
@@ -160,12 +160,12 @@ public class Job implements Serializable {
 	}
 
 	@JsonProperty("percentageComplete")
-	public Integer getPercentageComplete() {
+	public float getPercentageComplete() {
 		return percentageComplete;
 	}
 
 	@JsonProperty("percentageComplete")
-	public void setPercentageComplete(Integer percentageComplete) {
+	public void setPercentageComplete(float percentageComplete) {
 		this.percentageComplete = percentageComplete;
 	}
